+++ public/style.css
@@ -23,6 +23,14 @@
 ::-webkit-scrollbar-thumb:hover {
   background: #4c4e4eff;
 }
+::-webkit-color-swatch {
+  border: none;
+  border-radius: 5px;
+}
+::-moz-color-swatch {
+  border: none;
+  border-radius: 5px;
+}
 body {
   height: 100vh; width: 100vw;
   background: #1c1e1e;
@@ -39,69 +47,6 @@ body main {
   display: flex; justify-content: flex-start; align-items: flex-start; gap: 10px;
   overflow: auto;
   padding: 100px;
-}
-
-
-
-
-/* FORM */
-body #form-container {/* HEEPING THIS UNTIL ALL FORMS HAVE BEEN UDJUSTED */
-  background: #0009;
-  display: flex; justify-content: center; align-items: center;
-  position: fixed; left: 0px; top: 0px;
-  width: 100vw; height: 100vh;
-  backdrop-filter: blur(15px);
-  animation: blur-form-in 0.3s ease-in-out 1;
-}
-body #form {
-  /* display: flex; justify-content: center; align-items: center; gap: 5px;
-  background: #3338;
-  min-width: 300px;
-  padding: 10px;
-  border-radius: 10px;
-  box-shadow: 0 0 30px rgba(0, 0, 0, 0.6);
-  animation: pop-form-in 0.3s ease-in-out 1; */
-  /* border: 1px solid #3c3c3c; */
-}
-#form input,
-#form select {
-  height: 40px;
-  border-radius: 4px;
-  border: none;
-  font-size: 1em;
-  padding: 5px; padding-left: 10px;
-}
-#form input:focus,
-#form select:focus {
-  outline: none;
-}
-#form input:invalid {
-  background: #f99;
-}
-#form .tool-button {
-  min-width: 40px; height: 40px;
-  background-position: center; background-size: cover; background-repeat: no-repeat;
-  border: none;
-  box-sizing: border-box;
-  border-radius: 50%;
-  opacity: .6;
-  transition: all .1s ease-in-out;
-}
-#form #add-button {
-  background-image: url("./icons/plus-disabled.png");
-}
-#form #update-button {
-  background-image: url("./icons/check-disabled.png");
-}
-#form #add-button:enabled {
-  background-image: url("./icons/plus.png");
-}
-#form #update-button:enabled {
-  background-image: url("./icons/check.png");
-}
-#form .tool-button:hover {
-  opacity: 1;
-  cursor: pointer;
 }
